v012 update. big ass update! added a new menu with 2 items!
v012 update
This is going to be my workspace for today.
Running splash screen so that you dont have to wait as long before
something shows up on the screen.
edit: it does fucking nothing

added a settings screen
added device info screen (ip's and shit)
put the repair device there, delets the auth.json upon running. no
verification if you want to continue though LOL

reworked the main menu numbers, so it's a bit easier to shuffle things
around.
reworked the order of the menu.
most used functions in the front and back. i..e turn off all lights
slowly is on the front, and group control is on the back. group >
individual ights. and turn off all > turn on all
the 6 scenes are the only things that aren't very generic. the rest of
the code is pretty generic and should work on anyone's system.