scene transition time is now a thing!
v037
2017-02-04 2319 //57
bunch of tweaks here and there. biggest thing is that scene creation now
has adjustable transition time :D that's cool :)
thinking about making the menu system a generic function.... pass it a
bunch of things like names of functions to display shit or something
then give it a dict to build the menu from... idk...
to "start" this effort, i've made a generic/filler funtion called
binarydecision() which currently does nothing inttentionally
the idea behind this function is to put a "yes or no" decision anywhere
I want to without hard coding it in.
just give pass it a "question" in the form of a display_nlines function,
and it'll return whether or not the person
clicked yes or no. it wont be pretty, but it should be pretty
functional.