v041 update
v041
2017-02-11 2233 //57
binarydecision() is finished!
Pass it a function for the first thing, then a word or so for the
"answers"
We can now cancel scene creation!
Also added a user initiated update function! it'll go and pull the
latest file down from github and compare before installing! Then it'll
ask the user based on the binarydecision function!
Lots of goodies today!