Added settings-class
Only has timeformat now, and not implemented yet. Using Pickle to save
the settings to a file. Just add more local variables to the class to
add more settings. Do not use global variables, use local and get and
set-methods that saves the setting every time you change it so that you
dont have to remeber to do that manually