Added options for the quick-actions
Options for quick and longpress on the clock-screen. Would be really
cool to be able to select a specific scene, but that was too complicated
for now at least. Now you can toggle, turn off, on or do nothing.

Only tested on WSL without bridge. Seems like something is wierd with
the menu, but I cant find it