Added circuit diagrams
Created using Fritzing. Could not use the Pi Zero since it didn't want
to export. Also I couldn't find any encoder with a switch integrated, so
it isn't perfect. You can decide if you want to add it to the readme or
wiki, or not use it at all until it is fixed